renewal: by default, use a fraction of lifetime
Previously we defaulted to renewing at 30 days before expiry, and allowed users
to customize the config file to set a different value.

Instead, we should renew when 1/3 of the lifetime is left, or for shorter
certificates (<10 days), when 1/2 of the lifetime is left.

This still allows explicitly configured values to take precedence.