crypto_util_test: fix some bugs
The 2048-bit cert was referred to as `SS_CERT`, but the fact that
it was self-signed was not important to the tests, and is in fact
distracting for the test that verifies it is signed by a specific
chain. Rename the variables to `RSA2048_CERT_PATH` (the key length
is also not important to the tests but this is in keeping with
adjacent variables).

Some tests checked for "Example.com" with a capital "E" - change
this to plain "example.com" in line with the new testdata generation.

GetSANsFromCertTest::test_single was actually asserting it received
no SANs from `cert_512.pem`. This seems incorrect based on the test
name. The newly generated `cert_512.pem` has a single SAN; test for
that.

The sha256 test was checking for the hash of specific data. Use
different data and hardcode it into the test case.