renewal: use lineage-specific server for ARI
Previously, we were constructing an ACME client for ARI checking that used
the global value for `server`, not the one recorded in a lineage's renewal file.

This resulted in errors in the logs and failure to observe ARI for lineages that
used a non-default `--server` (e.g. staging or non-Let's Encrypt CAs).