acme.crypto_util: deprecate Format enum
This is a bit hacky. Since accessing enum variants doesn't invoke
the enum's constructor, we instead have to patch into the metaclass
__getattribute__ handler to warn users of the deprecated class.

This also replaces the default Format values in some now-deprecated
crypto_util functions with a None default, which is then checked at
function runtime. This is to avoid raising a ton of deprecation
warnings when those functions are declared.